// frontend/src/utils/interfaces/DashboardInterface.tsx
export interface DashboardProps {
    watchlistText: React.ReactNode;
    assetsText: React.ReactNode;
    view: React.ReactNode;
    window: React.ReactNode;
    watchlistCard: React.ReactNode;
    portfoliovalue: React.ReactNode;
    chart: React.ReactNode;
    graph: React.ReactNode;
    portfolio: React.ReactNode;
    coinText: React.ReactNode;
    info: React.ReactNode;
    infoText: React.ReactNode;
    bitcoinSlider: React.ReactNode;
    portfolioCardList: React.ReactNode;
    walletCard: React.ReactNode;
    transactions: React.ReactNode;
}